requirements.txt
chromadb==0.5.0
langchain==0.2.1
langchain-community==0.2.1
langchain-openai==0.1.7
openai==1.30.2
python-dotenv==1.0.1
tiktoken==0.7.0